Refactored utils folder, created getMailFromUid util
